Fixes 216: Restructure frontend components
