+++ src/components/ContentListTable/components/AddContent/AddContent.tsx
@@ -23,7 +23,7 @@ import { global_Color_200, global_link_Color } from '@patternfly/react-tokens';
 import { useFormik } from 'formik';
 import { useEffect, useMemo, useState } from 'react';
 import { createUseStyles } from 'react-jss';
-import Hide from '../Hide/Hide';
+import Hide from '../../../Hide/Hide';
 import {
   isValidURL,
   mapFormikToAPIValues,
@@ -32,18 +32,18 @@ import {
   magicURLList,
   FormikValues,
 } from './helpers';
-import useDebounce from '../../services/useDebounce';
+import useDebounce from '../../../../services/useDebounce';
 import ContentValidity from './components/ContentValidity';
 import {
   REPOSITORY_PARAMS_KEY,
   useAddContentQuery,
   useFetchGpgKey,
   useValidateContentList,
-} from '../../services/Content/ContentQueries';
-import { RepositoryParamsResponse } from '../../services/Content/ContentApi';
-import DropdownSelect from '../DropdownSelect/DropdownSelect';
+} from '../../../../services/Content/ContentQueries';
+import { RepositoryParamsResponse } from '../../../../services/Content/ContentApi';
+import DropdownSelect from '../../../DropdownSelect/DropdownSelect';
 import { useQueryClient } from 'react-query';
-import OptionalTooltip from '../OptionalTooltip/OptionalTooltip';
+import OptionalTooltip from '../../../OptionalTooltip/OptionalTooltip';
 
 interface Props {
   isLoading?: boolean;
@@ -575,7 +575,17 @@ const AddContent = ({ isLoading }: Props) => {
                               if (isValidURL(value)) {
                                 updateVariable(index, { gpgLoading: true });
                                 const gpgData = await fetchGpgKey(value);
-                                updateVariable(index, { gpgKey: gpgData, gpgLoading: false });
+                                updateVariable(index, {
+                                  gpgKey: gpgData,
+                                  gpgLoading: false,
+                                  ...(gpgKey === '' && !!value
+                                    ? {
+                                        metadataVerification:
+                                          !!validationList?.[index]?.url
+                                            ?.metadata_signature_present,
+                                      }
+                                    : {}),
+                                });
                               }
                             }}
                             onTextChange={(value) =>
